Added pagination for recipe listing for index.html.
An `iframe` is embedded in the index.html file which displays
by default `page-1.html`, containing a partial list of recipes.
Each page has a header bar with alphabetic links to jump to the
page with recipes starting with that letter (like a dictionary).
Classic page navigation buttons are also present at the bottom
of each page (e.g. `[1] [2] [4] [5]    [<<] [<] [3] [>] [>>]`).